Cast 64 to int for printf, portability.

+++ host2str.c
@@ -1927,7 +1927,7 @@ ldns_resolver_print(FILE *output, const ldns_resolver *r)
 	fprintf(output, "dnssec: %d\n", ldns_resolver_dnssec(r));
 	fprintf(output, "dnssec cd: %d\n", ldns_resolver_dnssec_cd(r));
 	fprintf(output, "trust anchors (%d listed):\n",
-		ldns_rr_list_rr_count(ldns_resolver_dnssec_anchors(r)));
+		(int)ldns_rr_list_rr_count(ldns_resolver_dnssec_anchors(r)));
 	ldns_rr_list_print(output, ldns_resolver_dnssec_anchors(r));
 	fprintf(output, "tsig: %s %s\n", ldns_resolver_tsig_keyname(r), ldns_resolver_tsig_algorithm(r));
 	fprintf(output, "debug: %d\n", ldns_resolver_debug(r));
@@ -1937,7 +1937,7 @@ ldns_resolver_print(FILE *output, const ldns_resolver *r)
 	fprintf(output, "\n");
 	fprintf(output, "apply default domain: %d\n", ldns_resolver_defnames(r));
 
-	fprintf(output, "searchlist (%d listed):\n",  ldns_resolver_searchlist_count(r));
+	fprintf(output, "searchlist (%d listed):\n",  (int)ldns_resolver_searchlist_count(r));
 	for (i = 0; i < ldns_resolver_searchlist_count(r); i++) {
 		fprintf(output, "\t");
 		ldns_rdf_print(output, s[i]);
@@ -1945,7 +1945,7 @@ ldns_resolver_print(FILE *output, const ldns_resolver *r)
 	}
 	fprintf(output, "apply search list: %d\n", ldns_resolver_dnsrch(r));
 
-	fprintf(output, "nameservers (%d listed):\n", ldns_resolver_nameserver_count(r));
+	fprintf(output, "nameservers (%d listed):\n", (int)ldns_resolver_nameserver_count(r));
 	for (i = 0; i < ldns_resolver_nameserver_count(r); i++) {
 		fprintf(output, "\t");
 		ldns_rdf_print(output, n[i]);
